arm64: mark x8 as used for struct return in memory
Before e0b94a3d6a, the assertion in spill() to check the registers
in use at the start of the block was skipped for the first block.
This opened up a class of bugs that are only detected with an extra
dummy block, such as [0].

In order for the assert to check registers used by the first block,
abi() now keeps track of the argument registers used by the function.
However, on arm64 this missed x8 when the function returned a struct
in caller-allocated memory.

[0] https://todo.sr.ht/~sircmpwn/hare/196